Revert "Index u16 -> u32 (#1787)"
This reverts commit 90bc90b0100834c558967c219d0be782cfe6a697.

+++ crates/lgn-graphics-renderer/src/resources/static_buffer.rs
@@ -93,7 +93,7 @@ impl UnifiedStaticBuffer {
         IndexBufferBinding {
             buffer: &self.buffer,
             byte_offset: 0,
-            index_type: IndexType::Uint32,
+            index_type: IndexType::Uint16,
         }
     }
 }
